refactor(config): the network types move out, and this is the move that failed the first time
Q-3, slice 2 and the last of it. `NetworkType`, `SyncDirection`, `VoteValue`, `VoteRoundType`, `NetworkMember`,
`VoteCast`, `VoteRound` and `NetworkConfig` move to `config/types-networks.ts`, re-exported from `types.ts` so no
importer changes and nothing behaves differently. Ratchet lowered 645 -> 578.

`config/types.ts` across the whole of Q-3: **677 -> 645 -> 578**. The gate's own comment said a fifth raise of that
file should be a split instead; this is the second half of it.

**This exact move was attempted before slice 1 and reverted, which is the interesting part.** `NetworkConfig`
references `SpaceMeta`, which belongs to spaces. With `types.ts` re-exporting the new file and the new file
importing `SpaceMeta` back from `types.ts`, that is a module cycle -- and TypeScript resolved it by degrading
`NetworkConfig` to `any`. `api/invite.ts` picked up three `TS7006` implicit-anys on its `members` callbacks while
every file in the diff compiled clean. It even reached a smaller 609 lines, which is exactly why the number was
never the thing to optimise.

So `SpaceMeta` went to `types-knowledge.ts` first -- a leaf that imports nothing, and cannot be half of a cycle.
This file imports it FROM the leaf. That ordering is the whole content of the two-slice plan.

Verified on the CALLERS, which is what the first attempt did not do:

    npm --prefix server run build 2>&1 | grep -c "TS7006"    # 0

The exported names are read off the moved block rather than hand-listed, because a hand list is how a re-export
quietly drops a type.

One gate failed and it was the more interesting failure of the two possible ones. `config-key-docs-coverage`
collects declared config fields from a hand-written list of source files, so when `NetworkConfig`'s fields moved it
stopped seeing them and reported `members` and `direction` as undocumented keys in the networks guide. A false
positive -- annoying, but the safe direction. Had the split gone the other way it would have checked LESS and
still passed, which is the same vacuous-gate shape this repo keeps finding.

Fixed by deriving the sources from the config directory (`types*.ts`) instead of listing them, so a further slice
is covered without touching the gate. Mutation-tested: narrowing it back to `types.ts` alone fails.
